+++ mlflow/tracking/_tracking_service/client.py
@@ -15,6 +15,7 @@ from mlflow.entities import (
     Param,
     RunStatus,
     RunTag,
+    TraceData,
     TraceRequestMetadata,
     TraceStatus,
     TraceTag,
@@ -25,7 +26,6 @@ from mlflow.exceptions import MlflowException
 from mlflow.protos.databricks_pb2 import INVALID_PARAMETER_VALUE, ErrorCode
 from mlflow.store.artifact.artifact_repository_registry import get_artifact_repository
 from mlflow.store.tracking import GET_METRIC_HISTORY_MAX_RESULTS, SEARCH_MAX_RESULTS_DEFAULT
-from mlflow.tracing.types.model import TraceData
 from mlflow.tracking._tracking_service import utils
 from mlflow.tracking.metric_value_conversion_utils import convert_metric_value_to_float_if_possible
 from mlflow.utils import chunk_list
