+++ tests/tracking/test_client.py
@@ -13,6 +13,8 @@ from mlflow.entities import (
     RunStatus,
     RunTag,
     SourceType,
+    SpanStatus,
+    SpanType,
     TraceRequestMetadata,
     TraceStatus,
     ViewType,
@@ -30,7 +32,6 @@ from mlflow.store.model_registry.sqlalchemy_store import (
 from mlflow.store.tracking import SEARCH_MAX_RESULTS_DEFAULT
 from mlflow.store.tracking.sqlalchemy_store import SqlAlchemyStore as SqlAlchemyTrackingStore
 from mlflow.tracing.types.constant import TraceMetadataKey
-from mlflow.tracing.types.model import SpanType, Status, StatusCode
 from mlflow.tracking import set_registry_uri
 from mlflow.tracking._model_registry import DEFAULT_AWAIT_MAX_SLEEP_SECONDS
 from mlflow.tracking._model_registry.utils import (
@@ -197,7 +198,9 @@ def test_start_and_end_trace(mock_trace_client):
             )
 
             res = self.square(z, trace_id, root_span.span_id)
-            self._client.end_trace(trace_id, outputs={"output": res}, status=Status(StatusCode.OK))
+            self._client.end_trace(
+                trace_id, outputs={"output": res}, status=SpanStatus(SpanStatus.StatusCode.OK)
+            )
             return res
 
         def square(self, t, trace_id, parent_span_id):
@@ -226,7 +229,7 @@ def test_start_and_end_trace(mock_trace_client):
     trace_info = traces[0].trace_info
     assert trace_info.request_id is not None
     assert trace_info.execution_time_ms >= 0.1 * 1e3  # at least 0.1 sec
-    assert trace_info.status.status_code == StatusCode.OK
+    assert trace_info.status == TraceStatus.OK
     trace_metadata_dict = {meta.key: meta.value for meta in trace_info.request_metadata}
     assert trace_metadata_dict[TraceMetadataKey.INPUTS] == '{"x": 1, "y": 2}'
     assert trace_metadata_dict[TraceMetadataKey.OUTPUTS] == '{"output": 25}'
@@ -260,12 +263,9 @@ def test_start_and_end_trace(mock_trace_client):
 
 def test_start_and_end_trace_before_all_span_end(mock_trace_client):
     # This test is to verify that the trace is still exported even if some spans are not ended
-    import mlflow
-    from mlflow.tracing.types.model import StatusCode
-
     class TestModel:
         def __init__(self):
-            self._client = mlflow.tracking.MlflowClient()
+            self._client = MlflowClient()
 
         def predict(self, x):
             root_span = self._client.start_trace(name="predict")
@@ -300,7 +300,7 @@ def test_start_and_end_trace_before_all_span_end(mock_trace_client):
     assert trace_info.request_id is not None
     assert trace_info.timestamp_ms is not None
     assert trace_info.execution_time_ms is not None
-    assert trace_info.status.status_code == StatusCode.OK
+    assert trace_info.status == TraceStatus.OK
 
     spans = traces[0].trace_data.spans
     assert len(spans) == 3  # The non-ended span should be also included in the trace
@@ -308,21 +308,21 @@ def test_start_and_end_trace_before_all_span_end(mock_trace_client):
     span_name_to_span = {span.name: span for span in spans}
     root_span = span_name_to_span["predict"]
     assert root_span.parent_span_id is None
-    assert root_span.status.status_code == StatusCode.OK
+    assert root_span.status.status_code == SpanStatus.StatusCode.OK
     assert root_span.start_time // 1e3 == trace_info.timestamp_ms
     assert (root_span.end_time - root_span.start_time) // 1e3 == trace_info.execution_time_ms
 
     ended_span = span_name_to_span["ended-span"]
     assert ended_span.parent_span_id == root_span.context.span_id
     assert ended_span.start_time < ended_span.end_time
-    assert ended_span.status.status_code == StatusCode.OK
+    assert ended_span.status.status_code == SpanStatus.StatusCode.OK
 
     # The non-ended span should have null end_time and UNSET status
     non_ended_span = span_name_to_span["non-ended-span"]
     assert non_ended_span.parent_span_id == root_span.context.span_id
     assert non_ended_span.start_time is not None
     assert non_ended_span.end_time is None
-    assert non_ended_span.status.status_code == StatusCode.UNSET
+    assert non_ended_span.status.status_code == SpanStatus.StatusCode.UNSET
 
 
 def test_start_span_raise_error_when_parent_span_id_is_not_provided():
